refactor: resolve circular dependency between dom.js and dom-dialogs.js
Remove backward-compatibility re-exports of createCustomModal, showPromptDialog,
showConfirmDialog from dom.js that created a circular dependency with dom-dialogs.js.
Updated all consumer imports to reference dom-dialogs.js directly.

Closes #163
