+++ src/components/flow-view.js
@@ -1,4 +1,5 @@
-import { _el, showPromptDialog, startInlineRename, renderButtonBar } from '../utils/dom.js';
+import { _el, startInlineRename, renderButtonBar } from '../utils/dom.js';
+import { showPromptDialog } from '../utils/dom-dialogs.js';
 import { generateId } from '../utils/id.js';
 import { registerComponent, getComponent } from '../utils/component-registry.js';
 import {
