+++ src/utils/tab-lifecycle.js
@@ -16,7 +16,8 @@
  */
 
 import { generateId } from './id.js';
-import { showConfirmDialog, _el } from './dom.js';
+import { _el } from './dom.js';
+import { showConfirmDialog } from './dom-dialogs.js';
 import { bus, EVENTS } from './events.js';
 import { WorkspaceTab } from './tab-manager-helpers.js';
 import { reattachLayout, syncFileTree } from './workspace-layout.js';
